imagiology/requirements.txt
Flask
Flask-SQLAlchemy
flask-migrate
flask-restful
itsdangerous
python-dotenv
